moved well specific aquifer lithology out of aquifers
