+++ app/backend/aquifers/models.py
@@ -159,26 +159,6 @@ class AquiferVulnerabilityCode(AuditModel):
         return '{} - {}'.format(self.code, self.description)
 
 
-class AquiferLithologyCode(AuditModel):
-    """
-    Choices for describing Completed Aquifer Lithology
-    """
-    code = models.CharField(primary_key=True, max_length=100, db_column='aquifer_lithology_code')
-    description = models.CharField(max_length=100)
-    display_order = models.PositiveIntegerField()
-
-    effective_date = models.DateTimeField(blank=True, null=True)
-    expiry_date = models.DateTimeField(blank=True, null=True)
-
-    class Meta:
-        db_table = 'aquifer_lithology_code'
-        ordering = ['display_order', 'code']
-        verbose_name_plural = 'Aquifer Lithology Codes'
-
-    def __str__(self):
-        return '{} - {}'.format(self.code, self.description)
-
-
 @reversion.register()
 class Aquifer(AuditModel):
     """
